scraper-dgb/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
selenium==4.15.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
lxml==4.9.3
requests==2.31.0